Upgrade from Chromium 104 to Chromium 105 (uplift to 1.43.x).

+++ browser/renderer_context_menu/brave_mock_render_view_context_menu.cc
@@ -133,7 +133,7 @@ void BraveMockRenderViewContextMenu::AddSubMenu(int command_id,
   item.has_submenu = true;
   items_.push_back(item);
 
-  for (int i = 0; i < model->GetItemCount(); ++i) {
+  for (size_t i = 0; i < model->GetItemCount(); ++i) {
     MockMenuItem sub_item;
     sub_item.is_submenu = true;
     if (model->GetTypeAt(i) != ui::MenuModel::TYPE_SEPARATOR) {
